feat: add auth module and infrastructure with expanded tests
Infrastructure:
- Add Docker support for containerized deployment
- Set up Alembic for database schema migrations and versioning
- Add Claude Code skills documentation for development practices

Authentication & Domain Refactoring:
- Refactor authentication from shared/auth to domain/auth module structure
- Add onboarding domain module for user initialization flows
- Implement auth controller, service, and schema layers
- Remove legacy shared auth and authorization modules

Testing Expansion:
- Add 2,900+ lines of test coverage gap analysis
- Expand e2e protocol, security, infrastructure, and ops test suites
- Update existing test suites for new authentication structure
- Add webhook, telemetry, email, and rate limit test coverage

This refactoring establishes proper infrastructure tooling, organizes
authentication into a dedicated domain module, and provides comprehensive
test coverage across the application stack.

+++ app/domain/payment/schemas.py
@@ -1,25 +1,23 @@
+from decimal import Decimal
 from uuid import UUID
 from datetime import datetime
 from pydantic import BaseModel
 from app.shared.base_domain.schemas import BaseSchemaResponse
 
 
-
 class SubscriptionTypeCreate(BaseModel):
     type: str
-    cost: float
+    cost: Decimal
 
 
 class SubscriptionTypeUpdate(BaseModel):
     type: str | None = None
-    cost: float | None = None
+    cost: Decimal | None = None
 
 
 class SubscriptionTypeResponse(BaseSchemaResponse):
     type: str
-    cost: float
-
-
+    cost: float  # serialized as float in JSON for client compatibility
 
 
 class UserServiceResponse(BaseSchemaResponse):
@@ -28,13 +26,11 @@ class UserServiceResponse(BaseSchemaResponse):
     is_active: bool
 
 
-
-
 class PaymentCreate(BaseModel):
     user_service_id: UUID
     subscription_type_id: UUID
     deposit_id: str
-    amount: float
+    amount: Decimal
 
 
 class PaymentResponse(BaseSchemaResponse):
@@ -43,11 +39,9 @@ class PaymentResponse(BaseSchemaResponse):
     expires_at: datetime
 
 
-
-
 class PaymentHistoryResponse(BaseSchemaResponse):
     payment_id: UUID
     deposit_id: str
-    amount: float
+    amount: float  # serialized as float in JSON for client compatibility
     period_start: datetime
     period_end: datetime
